Offer greedy & nondeterministic rule instances.
Note: This commit introduces a regression, since PB-rules no longer work. The
code is also a bit messy and opaque; I expect to clean it up in the future.

Description: It is as of now possible to define rules such that *at least one*
of the rules applicable to a node must lead to a closed subtableau.

Remember that a *greedy* heuristic is used to determine *which rule* is to be
chosen next, and *which formulas* the rule will consume. After all, we can
treat the *order* of rule application as irrelevant in terms of computability
(but not in terms of complexity).

However, we cannot take this greedy approach when the *productions* of the
rule can also take multiple possible values. Choosing a different production
will not lead to a simple permutation of the same tableau: instantiations
represent a set of *overlapping* rules, and making a choice will lead to a
tableau that is genuinely different.

Therefore, we must try all choices at this level, until we find one that leads
to closure. (If we could, we would try them all and only keep the shortest
one, but that is likely prohibitively expensive; nevertheless, we could still
offer an option to toggle shortest proof finding.)

To deal with this, I added a `compose` key to the system. This key determines
what to do with the different instantiations *generated* by the rule (it should
not affect the degenerative constraints of a rule, since a constraint only
limits the applicability of a rule but does not change the rule itself). We
will call this handler the *compositor*, and it can take the following values:

- `nondeterministic`: Default. Different instantiations simply produce
  different rules. Since the rules will overlap, this means that we may have
  to try out different paths until we find one that leads to a closed tableau.

- `greedy`: As it was before. When there are multiple instantiations of a rule,
  we will assume that the first applicable one is the one we need. (This should
  probably be the setting for PB-rules, as the order of application is
  irrelevant — they don't consume any formulas off the branch anyway.)