Simplify rule instantiation code.
Degenerative and generative constraints were previously interpreted in a
complicated way, since the logic was too general. They were presented in
a single YAML key, and the program had to figure out for itself which
part is meant to "generate" variable assignments and which part is meant
to "restrict" the application of the rule.

Moreover, rules were composed of a lot of different data types with
fancy infix constructors that looked great on the surface and improved
composability, but introduced a lot of pattern matching that hurt
readability.

These issues are now fixed:

- Users are now expected to supply restrictive constraints and
generative constraints in seperate keys. The code is greatly simplified.
This is a regression in that you can no longer supply different optional
pairs of generators & restrictions in a single rule, but such
functionality wasn't used and if it ever becomes an issue, it should be
easy to adapt.
- A single data type now holds (almost) all information relevant to a
rule.

As a consequence, some of the terminology has already shifted:

- Degenerative constraints are now 'constrainers' (ι-), whereas generative
  constraints are 'generators' (ι+).
- 'Proof's are now 'Tableau's, and a 'Tableau' is now a 'TableauSetting'.