Change data structure of Tableaux.
Instead of using a generic `Data.Tree` with custom `TableauNode`s, the
entire tableau is now custom.

This simplifies thinking about tableaux; partly because the datatype is
less convoluted, and partly because the type system will prevent such
nonsensical situations as a closed tableau with child nodes.

This paves the way for LaTeX output. The text printer is now also
prettier.

The downside is the loss of generality for, for example, the code
used to print tableau trees.