Fix & simplify renumbering function.
The tableau renumbering function was unnecessarily complex, because it
used the old numbers to construct the new numbers. The only thing that
was really necessary was to keep an association table.

Moreover, I think there was a mistake before, because the function did
not take into account the fact that a node may contain multiple
formulas.

Finally, the function now also takes an argument to indicate the
starting number.