Look for input files in standard directories.
- Relative paths to logic files are now also searched for in standard
data directories if they could not be found in the current directory.
The standard data directories are those in which Cabal installed the
data files, as well as ~/.local/share/judge on n*x systems or
%APPDATA% in Windows.
- Added an input file for the logic of proofs.

Note that it is not yet the case that extensions can be optionally left
out, nor that multiple files can be specified.