Reference contradicting formulas in closure.
Generated proofs now show the reference numbers of formulas that
triggered the closure of the branches. Closes issue #5.

There is a regression: negation is now *only* triggered by direct
conflict of marks F and T; the `negation` and `contradicts` functions
of `Logic.Judge.Formula` have been removed.

This is more consistent with the rest of the code, as it also expects
that the logical systems use signed formulas. However, at some point
this should be more streamlined and less hardcoded.