Define source directories and drop extensions.
Solves issue #2.

- Logic files in the data directories can be given without extension,
but must not be located in a subdirectory. This should give a 'native'
feel to shipped logic files (e.g. the user can do `judge LP` to derive
theorems of LP) without causing ambiguities (a path should really be an
absolute path or a path relative to the working directory — it is
counterintuitive if it can be relative to many different directories).
- A `--list-logics` option shows which logic files are available in the
data directories.
- The selected logical system is echoed back to the user.